+++ Dockerfile
@@ -8,6 +8,7 @@ WORKDIR /usr/src/app
 COPY . .
 
 RUN apk update && apk add git
+RUN pip3 install --upgrade pip
 RUN pip3 install --no-cache-dir -r requirements.txt
 RUN pip3 install -U tox
